+++ package.json
@@ -41,14 +41,6 @@
   "dependencies": {
     "@ant-design/icons": "^4.8.0",
     "@tanstack/react-query": "4.18.0",
-    "@tanstack/react-query-devtools": "4.18.0",
-    "@testing-library/jest-dom": "^5.16.5",
-    "@testing-library/user-event": "^14.4.3",
-    "@types/jest": "^29.5.1",
-    "@types/lodash.find": "^4.6.7",
-    "@types/lodash.isempty": "^4.4.7",
-    "@types/lodash.isequal": "^4.5.6",
-    "@types/lodash.orderby": "^4.6.7",
     "@vitest/coverage-c8": "^0.30.1",
     "antd": "5.0.7",
     "axios": "1.2.6",
@@ -83,7 +75,15 @@
     "@commitlint/cli": "17.3.0",
     "@commitlint/config-conventional": "17.3.0",
     "@rollup/plugin-alias": "4.0.3",
+    "@testing-library/jest-dom": "^5.16.5",
+    "@tanstack/react-query-devtools": "4.18.0",
     "@testing-library/react": "13.4.0",
+    "@testing-library/user-event": "^14.4.3",
+    "@types/jest": "^29.5.1",
+    "@types/lodash.find": "^4.6.7",
+    "@types/lodash.isempty": "^4.4.7",
+    "@types/lodash.isequal": "^4.5.6",
+    "@types/lodash.orderby": "^4.6.7",
     "@types/react": "18.0.27",
     "@types/react-dom": "18.0.10",
     "@types/react-query": "^1.2.9",
