+++ src/layout/index.ts
@@ -1,2 +1,1 @@
-export * from './page-content';
-export * from './navigation';
+export * from './main-layout';
