+++ src/layout/main-layout.tsx
@@ -1,13 +1,12 @@
 import { ToastContainer } from 'react-toastify';
-import { PageContent, Navigation } from '../layout';
+import { PageContent } from './page-content';
+import { Navigation } from './navigation';
 import 'react-toastify/dist/ReactToastify.css';
 
-const MainLayout = (): JSX.Element => (
+export const MainLayout = (): JSX.Element => (
   <>
     <Navigation />
     <PageContent />
     <ToastContainer position='bottom-left' />
   </>
 );
-
-export default MainLayout;
