+++ src/layout/page-content/index.ts
@@ -1,0 +1,1 @@
+export * from './page-content';
