+++ src/layout/page-content/page-content.tsx
@@ -2,9 +2,7 @@ import { Outlet } from 'react-router-dom';
 import { PageContentWrapper } from './page-content.styled';
 
 export const PageContent = (): JSX.Element => (
-  <>
-    <PageContentWrapper>
-      <Outlet />
-    </PageContentWrapper>
-  </>
+  <PageContentWrapper>
+    <Outlet />
+  </PageContentWrapper>
 );
